CodeFactor suggestions
