+++ DDEyC_API/Extensions/PartitionedCookieExtensions.cs
@@ -1,3 +1,4 @@
+using DDEyC_API.Middlewares;
 using DDEyC_API.Shared.Configuration;
 
 public static class PartitionedCookieExtensions
